Moved updateDeliveryInfo to a new file: userInfoHandling
