+++ ecosphere-goods/src/utilityFunctions/userAuth.js
@@ -43,13 +43,3 @@ export const fetchRole = async (uid) => {
         throw new Error('Error occurred while fetching role')
     }
 }
-
-export const updateDeliveryInfo = async (userID, address, country, zipCode, phoneNumber) => {
-    try {
-        const response = await axios.post('http://localhost:5000/auth/set-delivery-address', {
-            userID, address, country, zipCode, phoneNumber
-        })
-    } catch (error) {
-        throw new Error('Error while updating delivery address')
-    }
-}
